refactor(profiling): stack walking cleanup
 1. This renames the long-standing "stalk_walking" typo into
    "stack_walking" as it should have been.
 2. This moves run_time_cache cfg/cfg-nots into a detail mod.
